Platform/Marvell: ComPhyLib: Add missing SFI and RXAUI configuration
Despite the fact, that SFI and RXAUI modes are present on supported
feature list, their configuration was non existent and could not
be executed. This patch adds the missing initialization sequences.
Because ComPhySgmiiRFUPowerUp routine is common for SGMII, SFI and RXAUI,
rename it and reuse for those modes.

Contributed-under: TianoCore Contribution Agreement 1.0
Signed-off-by: Marcin Wojtas <mw@semihalf.com>